SUSHIGO.rar file
This file contains GUI that works with each screen as a different object from a different class. The PlayScreen contains the code that allows players to play, everything else kind of jsut sets up the game. Final score screen shows up after clicking the button after on the playscreen "final score". on the final score screen will be the final score for both players and who wins. Also there are two buttons on the bottom that will let a player either go back to the main menu, or hop back into a game of 1v1 sushi go.